System recode update
This system recode update replaces the old vues to the new jsons. It also adds a random theme selector, which will select a theme each load of the website.

Writing this update took a day and half to finish, due to lot of testing, especially working on the theme selector-